slime-clime.el: Stop breaking the keyboard and make kill/yang ~work
If you place the point directly on a CLIME image then keyboard
commands used to stop working (until you used the mouse to jump
somewhere else.) This is because the CLIME mouse-events keymap was
capturing all input events including the keyboard. This is fixed now!

Also: when you yank an image into a buffer, e.g. because you are
pasting your most important CLIM images into an *important* buffer for
future reference, Emacs tends to remove the keymap property. This
stops mouse-click-to-accept from working. I added a workaround for
this so that the keymap is added back to all CLIME images every time a
new ACCEPT request comes in.

There is a very long comment about all this in the source code now.