slime-clime.el: Optimize input context to O(n) instead of O(n^2)
Input context code was really slow. I have ~50K presentations in Emacs
and it would take a few seconds for ACCEPT to update all the images to
mark which ones should be enabled.

The profiler (`M-x profiler-start' and `M-x profiler-report') took me
straight to the problem i.e. that while already doing a linear sweep
of all presentations in all images, to decide which ones to enabled, I
was then also doing a linear search of the active presentation list.

Since I had about 50K presentations, and they were all of the same
type i.e. the all included in the list of acceptable IDs, it slowed
down a lot.

Now it's fast. I just switched the set of enabled IDs from a list to a
hashtable.